LayoutTests/editing/selection/ios/caret-rect-inside-scrollable-fixed-container-expected.txt

Verifies that when the scroll offset is extremely large, the caret rect for a focused text field inside a fixed container is positioned inside the text field

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS UIHelper.rectContainsOtherRect(textFieldRect, caretRect) is true
PASS successfullyParsed is true

TEST COMPLETE
